# perf(#8771): improve performance of task recalculation for many contacts