<!--
Please use semantic PR titles that respect this format:

<type>(#<issue number>): <subject>

Quick example:

feat(#1234): add hat wobble
^--^(#^--^): ^------------^
|     |      |
|     |      + - > subject
|     |
|     + -------- > issue number
|
+ -------------- > type: chore, feat, fix, perf.

https://docs.communityhealthtoolkit.org/contribute/code/workflow/#commit-message-format
-->

# Description

PouchDb is notoriously slow to query views with a large number of keys. This is what happened when task recalculation was triggered. For users with large numbers of contacts, these requests became so heavy it took minutes to complete, sometimes crashing PouchDb entirely. 

This adds a short-circuit to the key loaded query, when there are more than  500 keys, and instead gets all rows and filters them. 
The limit of 500 can be changed. In my sample test with ~10000 reports, it took roughly the same time to get all rows vs query with 500 keys. 

Performance difference: 
| contacts | reports | Query style                   | duration | duration | duration | duration | AVG    | IDB Crashes |
|------------| ------ |-------------------------------|----------|----------|----------|----------|--------|-------------|
| 6661       | 10000 | keys (current)                | 293s   | 276s  | 270s   | 293s   | 283s | 3           |
| 6661       | 10000 | all rows (proposed)           | 64s    | 63s    | 66s   | 58s    | 63s  | 0           |
| 6661       | 10000  | start_key + end_key (attempt) | 100s   | 95s    | 92s    | 92s    | 95s  | 0           |



#8771

# Code review checklist
<!-- Remove or comment out any items that do not apply to this PR; in the remaining boxes, replace the [ ] with [x]. -->
- [ ] Readable: Concise, well named, follows the [style guide](https://docs.communityhealthtoolkit.org/contribute/code/style-guide/), documented if necessary.
- [ ] Documented: Configuration and user documentation on [cht-docs](https://github.com/medic/cht-docs/)
- [ ] Tested: Unit and/or e2e where appropriate
- [ ] Internationalised: All user facing text
- [ ] Backwards compatible: Works with existing data and configuration or includes a migration. Any breaking changes documented in the release notes.

# Compose URLs
<!-- Do not change these!  CI will automatically update these to be the deep URLs -->
If Build CI hasn't passed, these may 404:

* [Core](https://staging.dev.medicmobile.org/_couch/builds_4/medic:medic:8771-skip-keys-param/docker-compose/cht-core.yml)
* [CouchDB Single](https://staging.dev.medicmobile.org/_couch/builds_4/medic:medic:8771-skip-keys-param/docker-compose/cht-couchdb.yml)
* [CouchDB Cluster](https://staging.dev.medicmobile.org/_couch/builds_4/medic:medic:8771-skip-keys-param/docker-compose/cht-couchdb-clustered.yml)

# License

The software is provided under AGPL-3.0. Contributions to this project are accepted under the same license.