@kennsippell 2023-12-27
13393 docs, 1015 contacts
Galaxy Tab A7 Lite (Production Hardware)

Branch | 1st | 2nd | 3rd
-- | -- | --
master | 57s | 68s | 72s
8771-skip-keys-param | 67s |  57s | 63s

Maybe not a huge win for CARES users ...