@dianabarsan 2023-12-27
Yea, users with 1000 contacts really isn't the target here (our query limit is 500), the target were deployments where users end up with 5000-6000 contacts.