@lorerod 2024-01-09
Hi @dianabarsan, sorry for the delay. I'm testing this locally. 
Instance up using [cht-couchdb.yml](https://staging.dev.medicmobile.org/_couch/builds_4/medic:medic:8771-skip-keys-param/docker-compose/cht-couchdb.yml) and [cht-core.yml](https://staging.dev.medicmobile.org/_couch/builds_4/medic:medic:8771-skip-keys-param/docker-compose/cht-core.yml). 

I first generated data using [this](https://github.com/medic/cht-core/tree/master/tests/scalability/generate-real-world-data) script, but I got errors when logging in with the created users.

So I decided to check with a simple data set and use the [scalability csv files](https://github.com/medic/cht-core/tree/master/tests/scalability/csv), and I'm getting the same error:

![image](https://github.com/medic/cht-core/assets/21312057/033d556e-904d-4439-8533-e14441b18725)

I will continue tomorrow to compare with the master branch and continue to try to make this work locally. It may be something on my local machine.