@lorerod 2024-01-10
Ok, finally, I was able to test this with a user with 6.000 contacts. I got a similar average: 69s on `rules-engine:tasks:all-contacts`
LGTM! Thanks!