@kennsippell 2024-01-11
I think the change is excellent. I don't have access to NSSD data for testing so I'm not sure what dataset to try this with. Sorry that the ~1000 contact dataset I tried wasn't very instructive.

![500](https://github.com/medic/cht-core/assets/9014751/44b4d729-c03d-46e7-8719-e4b60a39e5d7)